Feat: 폴더 구조 변경 및 내 정보보기 로그아웃 구현
